Changing int loop counters to size_t in EnvironmentOptions.cpp to remove compiler warnings from Bonn
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@503 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/GUI/EnvironmentOptions.cpp
@@ -147,7 +147,7 @@ EnvironmentOptions::DeleteObstacle(){
   vector<Model*>& sel = GetVizmo().GetSelectedModels();
   typedef vector<Model*>::iterator SI;
   for(SI i = sel.begin(); i!= sel.end(); i++){
-    for(int j=0;j<mBody.size();j++){
+    for(size_t j=0;j<mBody.size();j++){
       if(*i==mBody[j]&&!mBody[j]->IsActive())
         GetVizmo().GetEnv()->DeleteMBModel(mBody[j]);
     }
@@ -162,7 +162,7 @@ EnvironmentOptions::MoveObstacle(){
   vector<Model*>& sel = GetVizmo().GetSelectedModels();
   typedef vector<Model*>::iterator SI;
   for(SI i = sel.begin(); i!= sel.end(); i++){
-    for(int j=0;j<mBody.size();j++){
+    for(size_t j=0;j<mBody.size();j++){
       if(*i==mBody[j]&&!mBody[j]->IsActive()){
         ObstaclePosDialog o(mBody[j], m_mainWindow, this);
         o.exec();
